tmp/context_generator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "context-generator"
version = "0.1.0"
description = "Progressive context CLAUDE.md generator for AI/LLM workflows"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "pathspec>=0.11.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
treesitter = ["tree-sitter>=0.21.0", "tree-sitter-languages>=1.10.0"]
tiktoken = ["tiktoken>=0.5.0"]
ollama = ["ollama>=0.1.0"]
dev = ["pytest>=7.0.0", "ruff>=0.1.0"]

[project.scripts]
context-generator = "context_generator.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["context_generator"]

[tool.ruff]
line-length = 100
target-version = "py311"
